pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "codex-portfolio"
version = "0.1.0"
description = ""
authors = [
    {name = "CodexDLC",email = "primecodex@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "djlint (>=1.36.4,<2.0.0)",
    "django-modeltranslation (>=0.19.19,<0.20.0)",
    "django-htmx (>=1.27.0,<2.0.0)",
    "django-redis (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
